% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HYPERLIQUID_INTERVALS}
\alias{HYPERLIQUID_INTERVALS}
\title{Hyperliquid Candle Intervals}
\format{
A character vector of 14 interval codes.
}
\usage{
HYPERLIQUID_INTERVALS
}
\description{
The 14 candle intervals accepted by Hyperliquid's \code{candleSnapshot} endpoint,
in ascending order. Unlike some other venues there is no \verb{1s} or \verb{6h}
interval.
}
\examples{
HYPERLIQUID_INTERVALS

}
\keyword{datasets}
