% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{time_convert_from_hyperliquid}
\alias{time_convert_from_hyperliquid}
\title{Convert a Hyperliquid Timestamp to POSIXct}
\usage{
time_convert_from_hyperliquid(time_value)
}
\arguments{
\item{time_value}{Numeric; epoch milliseconds.}
}
\value{
POSIXct vector in UTC.
}
\description{
Hyperliquid returns every timestamp as epoch milliseconds. This converts that
form to a POSIXct in UTC.
}
\examples{
time_convert_from_hyperliquid(1700000000000)

}
