% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{time_convert_to_hyperliquid}
\alias{time_convert_to_hyperliquid}
\title{Convert a POSIXct to a Hyperliquid Timestamp}
\usage{
time_convert_to_hyperliquid(datetime)
}
\arguments{
\item{datetime}{(POSIXct) object(s) to convert.}
}
\value{
(numeric) whole-number epoch milliseconds.
}
\description{
Formats a POSIXct as the timestamp form Hyperliquid expects: whole-number
epoch milliseconds.
}
\examples{
dt <- lubridate::as_datetime("2023-11-14 22:13:20", tz = "UTC")
time_convert_to_hyperliquid(dt)

}
